+++ src/main/java/io/xlorey/fluxloader/Installer.java
@@ -33,6 +33,8 @@ public class Installer {
         add(new PatchLuaEventManager("zombie.Lua.LuaEventManager"));
         add(new PatchBanSteamIDCommand("zombie.commands.serverCommands.BanSteamIDCommand"));
         add(new PatchBanUserCommand("zombie.commands.serverCommands.BanUserCommand"));
+        add(new PatchUnbanUserCommand("zombie.commands.serverCommands.UnbanUserCommand"));
+        add(new PatchUnbanSteamIDCommand("zombie.commands.serverCommands.UnbanSteamIDCommand"));
         add(new PatchKickUserCommand("zombie.commands.serverCommands.KickUserCommand"));
         add(new PatchQuitCommand("zombie.commands.serverCommands.QuitCommand"));
         add(new PatchChatServer("zombie.network.chat.ChatServer"));