+++ src/main/java/io/xlorey/fluxloader/Installer.java
@@ -29,7 +29,8 @@ public class Installer {
      * of the IPatch interface, which encapsulates the logic for modifying specific game files.
      */
     private static final ArrayList<IPatch> patches = new ArrayList<>(){{
-        add(new PatchLuaManager("zombie.Lua.LuaManager$Exposer"));
+        add(new PatchLuaManager("zombie.Lua.LuaManager"));
+        add(new PatchLuaManagerExposer("zombie.Lua.LuaManager$Exposer"));
         add(new PatchGameWindow("zombie.GameWindow"));
         add(new PatchLuaEventManager("zombie.Lua.LuaEventManager"));
         add(new PatchBanSteamIDCommand("zombie.commands.serverCommands.BanSteamIDCommand"));