+++ src/main/java/io/xlorey/fluxloader/patches/PatchLuaManager.java
@@ -1,6 +1,5 @@
 package io.xlorey.fluxloader.patches;
 
-import io.xlorey.fluxloader.server.api.PlayerUtils;
 import io.xlorey.fluxloader.shared.EventManager;
 import io.xlorey.fluxloader.shared.LuaExposer;
 import io.xlorey.fluxloader.utils.PatchTools;
@@ -11,7 +10,7 @@ import javassist.expr.MethodCall;
 /**
  * Author: Deknil
  * GitHub: <a href=https://github.com/Deknil>https://github.com/Deknil</a>
- * Date: 27.02.2024
+ * Date: 29.02.2024
  * Description: LuaManager patcher
  * <p>FluxLoader © 2024. All rights reserved.</p>
  */
@@ -29,39 +28,28 @@ public class PatchLuaManager extends PatchFile{
      */
     @Override
     public void patch() {
-        PatchTools.patchMethod(getClassName(), "exposeAll", (ctClass, ctMethod) -> {
+        PatchTools.patchMethod(getClassName(), "LoadDirBase", "java.lang.String", (ctClass, ctMethod) -> {
             try {
                 String classCode =
                         "{" +
-                        "java.util.Set exposedClasses = " + LuaExposer.class.getName() + ".getExposedClasses();" +
-                        "java.util.Iterator iterator = exposedClasses.iterator();" +
-                        "while(iterator.hasNext()) {" +
-                        "java.lang.Class clazz = (java.lang.Class) iterator.next();" +
-                        "this.setExposed(clazz);" +
-                        "}" +
+                         EventManager.class.getName() + ".invokeEvent(\"onLuaFilesLoaded\", new Object[]{$1});" +
                         "}";
-                ctMethod.insertBefore(classCode);
-
-                ctMethod.instrument(new ExprEditor() {
-                    public void edit(MethodCall m) throws CannotCompileException {
-                        if (m.getMethodName().equals("exposeGlobalFunctions")) {
-                            String code =  "{" +
-                                        "java.util.Set exposedObjects = " + LuaExposer.class.getName() + ".getExposedGlobalObjects();" +
-                                        "java.util.Iterator iterator = exposedObjects.iterator();" +
-                                        "while(iterator.hasNext()) {" +
-                                        "java.lang.Object object = (java.lang.Object) iterator.next();" +
-                                        "this.exposeGlobalFunctions(object);" +
-                                        "}" +
-                                        "$proceed($$);" +
-                                        "}";
-                            m.replace(code);
-                        }
-                    }
-                });
+                ctMethod.insertAfter(classCode);
             } catch (CannotCompileException e) {
                 throw new RuntimeException(e);
             }
         });
 
+        PatchTools.patchMethod(getClassName(), "RunLuaInternal", (ctClass, ctMethod) -> {
+            try {
+                String classCode =
+                        "{" +
+                         EventManager.class.getName() + ".invokeEvent(\"onLuaScriptExecuted\", new Object[]{$1, $2});" +
+                        "}";
+                ctMethod.insertAfter(classCode);
+            } catch (CannotCompileException e) {
+                throw new RuntimeException(e);
+            }
+        });
     }
 }